[SPARK-XXXXX][UNSAFE] Fix MurmurHash3 on big endian platforms.
MurmurHash3 interprets each 4-byte data chunk as a little-endian
integer. This means that, since we are using unsafe operations, we
need to explicitly reverse the bytes in each chunk on big-endian
platforms. This change makes the output of the unsafe MurmurHash3
match the safe version on big-endian platforms.

I've left the hashInt and hashLong functions as-is for now. My
interpretation of these functions is that they perform the hash on
the integer value as if it were serialized in little-endian byte
order. Therefore no byte reversal is necessary.